+++ src/gui/gui2_entrylist.cpp
@@ -56,7 +56,7 @@ int GuiEntryList::addEntry(string name, string value)
 {
     entries.emplace_back(name, value);
     entriesChanged();
-    return entries.size() - 1;
+    return static_cast<int32_t>(entries.size() - 1);
 }
 
 int GuiEntryList::indexByValue(string value) const
@@ -81,7 +81,7 @@ void GuiEntryList::removeEntry(int index)
 
 int GuiEntryList::entryCount() const
 {
-    return entries.size();
+    return static_cast<int32_t>(entries.size());
 }
 
 string GuiEntryList::getEntryName(int index) const
