+++ src/repairCrew.cpp
@@ -161,7 +161,7 @@ void RepairCrew::update(float delta)
     if (position.x < -0.5)
     {
         ship->ship_template->interiorSize();
-        int n=irandom(0, ship->ship_template->rooms.size() - 1);
+        int n=irandom(0, static_cast<int32_t>(ship->ship_template->rooms.size() - 1));
         position = sf::Vector2f(ship->ship_template->rooms[n].position + sf::Vector2i(irandom(0, ship->ship_template->rooms[n].size.x - 1), irandom(0, ship->ship_template->rooms[n].size.y - 1)));
         target_position = sf::Vector2i(position);
     }
